refactor(parity): REFACTOR — extract per-action helper (T045, DR-11)
Tightens the parity gate now that GREEN established coverage.

- Move `setupFixture` / `teardownFixture` into `tests/parity-actions.ts`
  so the test file's `beforeEach`/`afterEach` collapse to a single line.
  The test file now expresses coverage, not lifecycle plumbing.
- Replace the `for (const spec of ACTION_TABLE)` loop with `it.each`,
  so vitest renders each generated case with a distinct name in the
  reporter (a parity break surfaces as `parity for action "cancel"`
  rather than being buried inside a single aggregate test).
- Reuse the `WORKFLOW_ACTIONS` tuple for the exhaustiveness sentinel
  instead of re-hardcoding the list.

No behavior change: the same 10 tests still run (9 actions +
exhaustiveness sentinel) and still pass.
